Find emulated memory region on Windows
There does not seem to be a nice abstract way to scan memory for all
platforms. For now we will reimplment the original windows algorithm
directly through FFI. We may want to try to find a specialized crate or
write our own in the future to avoid the unsafe use here.