Remove duplication of SharedPlayer on server side
Maintaining two SharedPlayers on the server (one in the lobby and one globally)
was extremely error-prone and didn't provide any real benefits, as none of the
shared player information matters without the context of a lobby.

The state's player map now maps between AuthIds and LobbyIds. Additionally the
AuthId is now being used as a UID for players within a lobby. This will need
to be modified eventually to keep the player's AuthId and UID separate.